Crate sos_alloc

SOS memory allocation library

This is in its own crate so it can be used by kernel-space and user-space OS components.

We reproduce the Rust Allocator trait described in Rust RFC 1398, so that our allocators can implement it. Once RFC 1398 is implemented, we will be able to remove the Allocator and Layout types from this crate.

For more information on the Allocator API, refer to: + the text of RFC 1398 + rust-lang/rust tracking issue

Reexports

pub use frame::Allocator as FrameAllocator;

pub use frame::Lender as FrameLender;

pub use system::*;

Modules

borrow

buddy
Simple buddy-block allocator

bump_ptr

frame
Frame allocation

system

Structs

CannotReallocInPlace
The CannotReallocInPlace error is used when fn realloc_in_place was unable to reuse the given memory block for a requested layout.

Excess
Represents the combination of a starting address and a total capacity of the returned block.

Layout
Category for a memory record.

Enums

AllocErr
The AllocErr error specifies whether an allocation failure is specifically due to resource exhaustion or if it is due to something wrong when combining the given input arguments with this allocator.

Traits

Allocator
An implementation of Allocator can allocate, reallocate, and deallocate arbitrary blocks of data described via Layout.

Type Definitions

Address

Alignment

AllocResult

Capacity

Size