Module sos_alloc::buddy

Simple buddy-block allocator

Reexports

pub use self::system::BuddyFrameAllocator;

Modules

system
This module integrates the buddy heap allocator into the Rust runtime.

Structs

FreeBlock
A free block header stores a pointer to the next and previous free blocks.

Heap
Structure with data for implementing the buddy block allocation strategy.

Type Definitions

FreeList
A FreeList is a list of unique free blocks