Module sos_alloc::buddy::system

This module integrates the buddy heap allocator into the Rust runtime.

Structs

BuddyFrameAllocator
A frame allocator using the system's buddy-block heap allocator

Constants

NUM_FREE_LISTS
The number of free lists for the kernel heap

Functions

__rust_allocate

__rust_deallocate

__rust_reallocate

__rust_reallocate_inplace
This is currently unsupported, so we just silently ignore it and return the old size.

__rust_usable_size

init_heap⚠
Initialize the system heap at the given start address