Trait sos_alloc::Allocator

pub unsafe trait Allocator {
    unsafe fn alloc(&mut self, layout: Layout) -> Result<Address, AllocErr>;
    unsafe fn dealloc(&mut self, ptr: Address, layout: Layout);

    fn oom(&mut self, _: AllocErr) -> ! { ... }
    unsafe fn usable_size(&self, layout: &Layout) -> (Capacity, Capacity) { ... }
    unsafe fn realloc(
        &mut self, 
        ptr: Address, 
        layout: Layout, 
        new_layout: Layout
    ) -> Result<Address, AllocErr> { ... }
    unsafe fn alloc_excess(
        &mut self, 
        layout: Layout
    ) -> Result<Excess, AllocErr> { ... }
    unsafe fn realloc_excess(
        &mut self, 
        ptr: Address, 
        layout: Layout, 
        new_layout: Layout
    ) -> Result<Excess, AllocErr> { ... }
    unsafe fn realloc_in_place(
        &mut self, 
        ptr: Address, 
        layout: Layout, 
        new_layout: Layout
    ) -> Result<(), CannotReallocInPlace> { ... }
    unsafe fn alloc_one<T>(&mut self) -> Result<Unique<T>, AllocErr>
    where
        Self: Sized,
    { ... }
    unsafe fn dealloc_one<T>(&mut self, ptr: Unique<T>)
    where
        Self: Sized,
    { ... }
    unsafe fn alloc_array<T>(&mut self, n: usize) -> Result<Unique<T>, AllocErr>
    where
        Self: Sized,
    { ... }
    unsafe fn realloc_array<T>(
        &mut self, 
        ptr: Unique<T>, 
        n_old: usize, 
        n_new: usize
    ) -> Result<Unique<T>, AllocErr>
    where
        Self: Sized,
    { ... }
    unsafe fn dealloc_array<T>(
        &mut self, 
        ptr: Unique<T>, 
        n: usize
    ) -> Result<(), AllocErr>
    where
        Self: Sized,
    { ... }
    unsafe fn alloc_unchecked(&mut self, layout: Layout) -> Option<Address> { ... }
    unsafe fn realloc_unchecked(
        &mut self, 
        ptr: Address, 
        layout: Layout, 
        new_layout: Layout
    ) -> Option<Address> { ... }
    unsafe fn alloc_excess_unchecked(
        &mut self, 
        layout: Layout
    ) -> Option<Excess> { ... }
    unsafe fn realloc_excess_unchecked(
        &mut self, 
        ptr: Address, 
        layout: Layout, 
        new_layout: Layout
    ) -> Option<Excess> { ... }
    unsafe fn alloc_array_unchecked<T>(&mut self, n: usize) -> Option<Unique<T>>
    where
        Self: Sized,
    { ... }
    unsafe fn realloc_array_unchecked<T>(
        &mut self, 
        ptr: Unique<T>, 
        n_old: usize, 
        n_new: usize
    ) -> Option<Unique<T>>
    where
        Self: Sized,
    { ... }
    unsafe fn dealloc_array_unchecked<T>(&mut self, ptr: Unique<T>, n: usize)
    where
        Self: Sized,
    { ... }
}

An implementation of Allocator can allocate, reallocate, and deallocate arbitrary blocks of data described via Layout.

Some of the methods require that a layout fit a memory block. What it means for a layout to "fit" a memory block means is that the following two conditions must hold:

The block's starting address must be aligned to layout.align().

The block's size must fall in the range [use_min, use_max], where:

use_min is self.usable_size(layout).0, and

use_max is the capacity that was (or would have been) returned when (if) the block was allocated via a call to alloc_excess or realloc_excess.

Note that:

the size of the layout most recently used to allocate the block is guaranteed to be in the range [use_min, use_max], and

a lower-bound on use_max can be safely approximated by a call to usable_size.

Required Methods

unsafe fn alloc(&mut self, layout: Layout) -> Result<Address, AllocErr>

Returns a pointer suitable for holding data described by layout, meeting its size and alignment guarantees.

The returned block of storage may or may not have its contents initialized. (Extension subtraits might restrict this behavior, e.g. to ensure initialization.)

Returning Err indicates that either memory is exhausted or layout does not meet allocator's size or alignment constraints.

Implementations are encouraged to return Err on memory exhaustion rather than panicking or aborting, but this is not a strict requirement. (Specifically: it is legal to use this trait to wrap an underlying native allocation library that aborts on memory exhaustion.)

unsafe fn dealloc(&mut self, ptr: Address, layout: Layout)

Deallocate the memory referenced by ptr.

ptr must have previously been provided via this allocator, and layout must fit the provided block (see above); otherwise yields undefined behavior.

Provided Methods

fn oom(&mut self, _: AllocErr) -> !

Allocator-specific method for signalling an out-of-memory condition.

Implementations of the oom method are discouraged from infinitely regressing in nested calls to oom. In practice this means implementors should eschew allocating, especially from self (directly or indirectly).

Implementions of this trait's allocation methods are discouraged from panicking (or aborting) in the event of memory exhaustion; instead they should return an appropriate error from the invoked method, and let the client decide whether to invoke this oom method.

unsafe fn usable_size(&self, layout: &Layout) -> (Capacity, Capacity)

Returns bounds on the guaranteed usable size of a successful allocation created with the specified layout.

In particular, for a given layout k, if usable_size(k) returns (l, m), then one can use a block of layout k as if it has any size in the range [l, m] (inclusive).

(All implementors of fn usable_size must ensure that l <= k.size() <= m)

Both the lower- and upper-bounds (l and m respectively) are provided: An allocator based on size classes could misbehave if one attempts to deallocate a block without providing a correct value for its size (i.e., one within the range [l, m]).

Clients who wish to make use of excess capacity are encouraged to use the alloc_excess and realloc_excess instead, as this method is constrained to conservatively report a value less than or equal to the minimum capacity for all possible calls to those methods.

However, for clients that do not wish to track the capacity returned by alloc_excess locally, this method is likely to produce useful results.

unsafe fn realloc(
&mut self,
ptr: Address,
layout: Layout,
new_layout: Layout
) -> Result<Address, AllocErr>

Returns a pointer suitable for holding data described by new_layout, meeting its size and alignment guarantees. To accomplish this, this may extend or shrink the allocation referenced by ptr to fit new_layout.

ptr must have previously been provided via this allocator.

layout must fit the ptr (see above). (The new_layout argument need not fit it.)

Behavior undefined if either of latter two constraints are unmet.

In addition, new_layout should not impose a different alignment constraint than layout. (In other words, new_layout.align() should equal layout.align().) However, behavior is well-defined (though underspecified) when this constraint is violated; further discussion below.

If this returns Ok, then ownership of the memory block referenced by ptr has been transferred to this allocator. The memory may or may not have been freed, and should be considered unusable (unless of course it was transferred back to the caller again via the return value of this method).

Returns Err only if new_layout does not meet the allocator's size and alignment constraints of the allocator or the alignment of layout, or if reallocation otherwise fails. (Note that did not say "if and only if" -- in particular, an implementation of this method can return Ok if new_layout.align() != old_layout.align(); or it can return Err in that scenario, depending on whether this allocator can dynamically adjust the alignment constraint for the block.)

If this method returns Err, then ownership of the memory block has not been transferred to this allocator, and the contents of the memory block are unaltered.

unsafe fn alloc_excess(&mut self, layout: Layout) -> Result<Excess, AllocErr>

Behaves like fn alloc, but also returns the whole size of the returned block. For some layout inputs, like arrays, this may include extra storage usable for additional data.

unsafe fn realloc_excess(
&mut self,
ptr: Address,
layout: Layout,
new_layout: Layout
) -> Result<Excess, AllocErr>

Behaves like fn realloc, but also returns the whole size of the returned block. For some layout inputs, like arrays, this may include extra storage usable for additional data.

unsafe fn realloc_in_place(
&mut self,
ptr: Address,
layout: Layout,
new_layout: Layout
) -> Result<(), CannotReallocInPlace>

Attempts to extend the allocation referenced by ptr to fit new_layout.

ptr must have previously been provided via this allocator.

layout must fit the ptr (see above). (The new_layout argument need not fit it.)

Behavior undefined if either of latter two constraints are unmet.

If this returns Ok, then the allocator has asserted that the memory block referenced by ptr now fits new_layout, and thus can be used to carry data of that layout. (The allocator is allowed to expend effort to accomplish this, such as extending the memory block to include successor blocks, or virtual memory tricks.)

If this returns Err, then the allocator has made no assertion about whether the memory block referenced by ptr can or cannot fit new_layout.

In either case, ownership of the memory block referenced by ptr has not been transferred, and the contents of the memory block are unaltered.

unsafe fn alloc_one<T>(&mut self) -> Result<Unique<T>, AllocErr> where
Self: Sized,

Allocates a block suitable for holding an instance of T.

Captures a common usage pattern for allocators.

The returned block is suitable for passing to the alloc/realloc methods of this allocator.

May return Err for zero-sized T.

unsafe fn dealloc_one<T>(&mut self, ptr: Unique<T>) where
Self: Sized,

Deallocates a block suitable for holding an instance of T.

The given block must have been produced by this allocator, and must be suitable for storing a T (in terms of alignment as well as minimum and maximum size); otherwise yields undefined behavior.

Captures a common usage pattern for allocators.

unsafe fn alloc_array<T>(&mut self, n: usize) -> Result<Unique<T>, AllocErr> where
Self: Sized,

Allocates a block suitable for holding n instances of T.

Captures a common usage pattern for allocators.

The returned block is suitable for passing to the alloc/realloc methods of this allocator.

May return Err for zero-sized T or n == 0.

Always returns Err on arithmetic overflow.

unsafe fn realloc_array<T>(
&mut self,
ptr: Unique<T>,
n_old: usize,
n_new: usize
) -> Result<Unique<T>, AllocErr> where
Self: Sized,

Reallocates a block previously suitable for holding n_old instances of T, returning a block suitable for holding n_new instances of T.

Captures a common usage pattern for allocators.

The returned block is suitable for passing to the alloc/realloc methods of this allocator.

May return Err for zero-sized T or n == 0.

Always returns Err on arithmetic overflow.

unsafe fn dealloc_array<T>(
&mut self,
ptr: Unique<T>,
n: usize
) -> Result<(), AllocErr> where
Self: Sized,

Deallocates a block suitable for holding n instances of T.

Captures a common usage pattern for allocators.

unsafe fn alloc_unchecked(&mut self, layout: Layout) -> Option<Address>

Returns a pointer suitable for holding data described by layout, meeting its size and alignment guarantees.

The returned block of storage may or may not have its contents initialized. (Extension subtraits might restrict this behavior, e.g. to ensure initialization.)

Returns None if request unsatisfied.

Behavior undefined if input does not meet size or alignment constraints of this allocator.

unsafe fn realloc_unchecked(
&mut self,
ptr: Address,
layout: Layout,
new_layout: Layout
) -> Option<Address>

Returns a pointer suitable for holding data described by new_layout, meeting its size and alignment guarantees. To accomplish this, may extend or shrink the allocation referenced by ptr to fit new_layout. (In other words, ownership of the memory block associated with ptr is first transferred back to this allocator, but the same block may or may not be transferred back as the result of this call.)

ptr must have previously been provided via this allocator.

layout must fit the ptr (see above). (The new_layout argument need not fit it.)

new_layout must meet the allocator's size and alignment constraints. In addition, new_layout.align() must equal layout.align(). (Note that this is a stronger constraint that that imposed by fn realloc.)

Behavior undefined if any of latter three constraints are unmet.

If this returns Some, then the memory block referenced by ptr may have been freed and should be considered unusable.

Returns None if reallocation fails; in this scenario, the original memory block referenced by ptr is unaltered.

unsafe fn alloc_excess_unchecked(&mut self, layout: Layout) -> Option<Excess>

Behaves like fn alloc_unchecked, but also returns the whole size of the returned block.

unsafe fn realloc_excess_unchecked(
&mut self,
ptr: Address,
layout: Layout,
new_layout: Layout
) -> Option<Excess>

Behaves like fn realloc_unchecked, but also returns the whole size of the returned block.

unsafe fn alloc_array_unchecked<T>(&mut self, n: usize) -> Option<Unique<T>> where
Self: Sized,

Allocates a block suitable for holding n instances of T.

Captures a common usage pattern for allocators.

Requires inputs are non-zero and do not cause arithmetic overflow, and T is not zero sized; otherwise yields undefined behavior.

unsafe fn realloc_array_unchecked<T>(
&mut self,
ptr: Unique<T>,
n_old: usize,
n_new: usize
) -> Option<Unique<T>> where
Self: Sized,

Reallocates a block suitable for holding n_old instances of T, returning a block suitable for holding n_new instances of T.

Captures a common usage pattern for allocators.

Requires inputs are non-zero and do not cause arithmetic overflow, and T is not zero sized; otherwise yields undefined behavior.

unsafe fn dealloc_array_unchecked<T>(&mut self, ptr: Unique<T>, n: usize) where
Self: Sized,

Deallocates a block suitable for holding n instances of T.

Captures a common usage pattern for allocators.

Requires inputs are non-zero and do not cause arithmetic overflow, and T is not zero sized; otherwise yields undefined behavior.